requirements.txt
streamlit
pandas
keepa
requests
google-generativeai
